Restructure conftest to place fixtures in the appropriate test group
