+++ test/test_commands/test_event/test_next_week.py
@@ -2,7 +2,7 @@ import datetime
 
 import discord
 import pytest
-from conftest import WEEK_EMBEDS
+from assets import WEEK_EMBEDS
 from freezegun import freeze_time
 from helper.helper_functions import check_payload_equal
